TaskApiRest/requirements.txt
fastapi
uvicorn[standard]
pydantic[dotenv]
redis
zeep
httpx
python-jose[cryptography]
passlib[bcrypt]
pydantic-settings
pytest
grpcio
grpcio-tools
